requirements.txt
# espnet==0.10.6
# ignite==1.1.0
# pytorch-ignite==0.4.7
# jams==0.3.4
librosa==0.8.0
matplotlib==3.1.2
numpy==1.18
numba==0.48
pandas==0.25.3
# pretty_midi==0.2.9
PyYAML==6.0
# pytorch-lightning==1.5.10
scikit_learn==1.0.2
scipy==1.4.1
seaborn==0.11.1
# SoundFile==0.10.3.post1
# sox==1.4.1
# tensorboardX==2.5
# torch==1.9.0
# torch_lr_finder==0.2.1
# torch-optimizer==0.3.0
tqdm==4.63.0
